api/python/requirements_dev.txt
torch==1.10.1
bokeh==2.4.2
pandas==1.3.5
seaborn==0.11.2
matplotlib==3.5.1
pytest==6.2.5
scikit-learn==1.0.1
umap-learn==0.5.2
